+++ example/lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:encrypt_decrypt_plus/cipher/cipher.dart';
+import 'package:encrypt_decrypt_plus/cipher/aes.dart';
 import 'package:flutter/material.dart';
 
 void main() {
@@ -16,7 +16,8 @@ class _MyAppState extends State<MyApp> {
   final String _plainTxt = "Hello datadirr";
   String _encryptTxt = "";
   String _decryptTxt = "";
-  final Cipher _cipher = Cipher(secretKey: "datadirr");
+  final AES _aes = AES(
+      secretKey: "datadirrdatadirrdatadirrdatadirr", iv: "datadirrdatadirr");
 
   @override
   Widget build(BuildContext context) {
@@ -30,13 +31,13 @@ class _MyAppState extends State<MyApp> {
                 Text("plainTxt: $_plainTxt"),
                 ElevatedButton(
                     onPressed: () {
-                      _encrypt();
+                      _encryptAES256CBC();
                     },
                     child: const Text("Encrypt")),
                 Text("encryptTxt: $_encryptTxt"),
                 ElevatedButton(
                     onPressed: () {
-                      _decrypt();
+                      _decryptAES256CBC();
                     },
                     child: const Text("Decrypt")),
                 Text("decryptTxt: $_decryptTxt"),
@@ -48,13 +49,13 @@ class _MyAppState extends State<MyApp> {
     );
   }
 
-  void _encrypt() async {
-    _encryptTxt = _cipher.xorEncode(_plainTxt);
+  void _encryptAES256CBC() async {
+    _encryptTxt = _aes.encryptAES256CBC(_plainTxt);
     setState(() {});
   }
 
-  void _decrypt() async {
-    _decryptTxt = _cipher.xorDecode(_encryptTxt);
+  void _decryptAES256CBC() async {
+    _decryptTxt = _aes.decryptAES256CBC(_encryptTxt);
     setState(() {});
   }
 }